Create last_round.sma
It's annoying when you are playing and suddenly you hear the map change countdown in the middle of the round and than the map changes...
Not anymore!

This is a simple plugin that changes the map after the last round is played. It's useful if you don't have it included in the change map plugin or you are running a classic server or a server with random maps. (no mapvote)
Any plugins that extend the map will break it's function for sure! (didn't test it, but theoretically it will)